+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.1.2 (2020-12-23)
+  * 修改公钥和私钥为空判断
+  * 添加了新方法名 rsaSign， 功能同 SignStrMessage
+  * 添加了新方法名 rsaCheck， 功能同 VerifyStrMessage
+
 ### 0.1.1 (2017-11-7)
 
   * Added IntrospectionProcessor to get info from where the logger was called